Move migration for apps to 1.3.3 all together (#15944)
(cherry picked from commit d1d964867dede9ccedc7c5960d0c96175d17c330)
