+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v133/Migration.java
@@ -1,4 +1,4 @@
-package org.openmetadata.service.migration.mysql.v131;
+package org.openmetadata.service.migration.mysql.v133;
 
 import static org.openmetadata.service.migration.utils.v131.MigrationUtil.migrateCronExpression;
